Use an HTTP connection to Docker
Instead of just calling Docker commands, this commit causes the app to use a
faster and more reliable way of directly connecting to the Docker daemon and
sending commands to it instead of through the CLI. This allows us to easily
see what is happening when an image is created.

This commit also adds a 5s timeout to any running containers, preventing CPU
maxing out by students for a prolonged period of time.